+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 + [x] Renamed `radio-button` to `toggle-button`
 + [x] Fixed attribute `data-caption` for `radio` component
-+ [x] Added component `cutter`
++ [x] Added component `cutter` for visual cutting page
++ [x] Added new events to `table` component: `onDrawHead`, `onDrawFoot`, `onDrawBody`, `onInspectorOpen`, `onInspectorClose`, `onFieldShow`, `onFieldHide`, `onFieldToggle`, `onFieldMove`, `onSort`, `onPageChange`, `onRowClick`, `onRowDblClick`, `onCellClick`, `onCellDblClick`